Fix Signature Verification issues

+++ lib/ami_middleware/crypto/rsa_aspeed.h
@@ -21,6 +21,6 @@ void rsa_engine_function_test(void);    // rsa functions testing
 #endif
 
 int decrypt_aspeed (const struct rsa_key *key, const uint8_t *encrypted, size_t in_length, uint8_t *decrypted, size_t out_length);
-int sig_verify_aspeed (const struct rsa_key *key, const uint8_t *signature, size_t sig_length, const uint8_t *match, size_t match_length);
+int sig_verify_aspeed (const struct rsa_key *key, const uint8_t *signature, int sig_length, const uint8_t *match, size_t match_length);
 int rsa_sig_verify_test(void);
 #endif	/* ZEPHYR_INCLUDE_RSA_API_MIDLEYER_H_ */
